requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
cryptography>=41.0.0  # For API key encryption
Flask-Limiter==3.5.0
Flask-SocketIO==5.3.6
flasgger==0.9.7.1
flask-talisman==1.1.0
Flask-Caching==2.1.0
prometheus-flask-exporter==0.23.0
python-json-logger==2.0.7
psycopg2-binary==2.9.7
Markdown==3.5.1
python-dotenv>=1.0.1
marshmallow==3.20.1
email-validator==2.1.0
bleach==6.1.0
Pillow==10.1.0
WeasyPrint==60.2  # Commented out due to system dependencies
python-docx==1.1.0
gunicorn==21.2.0
gevent==23.9.1
pytest==7.4.2
pytest-flask==1.3.0
mypy==1.14.1
ruff==0.14.9
types-bleach==6.3.0.20251115
types-Flask-Cors==6.0.0.20250809
types-Flask-Migrate==4.1.0.20250809
types-Flask-SocketIO==5.5.0.20250809
types-Markdown==3.10.0.20251106
types-Pillow==10.2.0.20240822
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
konlpy==0.6.0
python-mecab-ko
mecab-python3==1.0.6
opensearch-py==3.0.0
orgparse==0.3.2
charset-normalizer==3.3.2
PyYAML==6.0.1
requests>=2.31.0
openai==1.3.0
anthropic>=0.18.0
pytesseract==0.3.10
PyMuPDF==1.23.14
google-cloud-vision==3.4.5
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
numpy==1.26.2
markitdown[all]==0.1.2
